Fixed a hardfault, keil warnings, and began sending data to exosite
The hardfault was caused by a unaligned data access in a specific region
of memory.  This is what the file system is now not allowed to do word
accesses.  The file system was not enforcing aligned accesses.  The
cortex M4 normally doesn't mind unaligned accesses, except in a few
regions of memory.